+++ lib/src/repository/extra_files_repository.dart
@@ -1,4 +1,7 @@
+import 'dart:io';
+
 import 'package:snaply/src/data_holders/custom_attributes_holder.dart';
+import 'package:snaply/src/data_holders/custom_files_holder.dart';
 import 'package:snaply/src/entities/report_file.dart';
 import 'package:snaply/src/logger/snaply_logger.dart';
 import 'package:snaply/src/platform_interface/snaply_platform_interface.dart';
@@ -8,12 +11,15 @@ class ExtraFilesRepository {
     required SnaplyPlatformInterface platform,
     required SnaplyLogger logger,
     required CustomAttributesHolder customAttributesHolder,
+    required CustomFilesHolder customFilesHolder,
   })  : _platform = platform,
         _logger = logger,
-        _customAttributesHolder = customAttributesHolder;
+        _customAttributesHolder = customAttributesHolder,
+        _customFilesHolder = customFilesHolder;
   final SnaplyPlatformInterface _platform;
   final SnaplyLogger _logger;
   final CustomAttributesHolder _customAttributesHolder;
+  final CustomFilesHolder _customFilesHolder;
 
   Future<List<ReportFile>> getExtraFiles({
     required Map<String, Map<String, String>> reportAttrs,
@@ -30,9 +36,21 @@ class ExtraFilesRepository {
     );
     final logsFile = LogsFile(logs: _logger.logs);
 
+    final customFiles = _customFilesHolder.customFiles.entries
+        .map((entry) => File(entry.value))
+        .map(
+          (file) => CustomFile(
+            filePath: file.path,
+            exists: file.existsSync(),
+            length: file.lengthSync(),
+          ),
+        )
+        .toList();
+
     return [
       logsFile,
       attrsFile,
+      ...customFiles,
     ];
   }
 }